libdrm: Fix bogus increment of a property set object count
If drmModePropertySetAdd() is called with the following property pairs
it will mistakenly increase the object count:

  (4, 1), (5,1), (4, 2).

When adding the third pair, the new_obj variable would be set when prev
points to (4, 1) since it disregarded the value of the previous item.

This patch changes the logic so that the obj count is incresed only if
both the previous and next items have differing object ids.